FIX: lidar_map_factory: all pointclouds from both lidars are used. ADD: support of 2 lidars

+++ slam/frontend_manager/edge_factories/imu_preintegration/utils.py
@@ -37,7 +37,7 @@ def get_previous_vertex(
     storage: VertexStorage,
     timestamp: int,
     index: int,
-    time_margin: float,
+    time_margin: int,
 ) -> GraphVertex:
     """Gets previous vertex if found or creates a new one.
